requirements.txt
flask
python-telegram-bot
requests
gunicorn
asyncio  # Добавьте эту строку